refactor: update module structure and S3 configuration for internal package usage
