+++ k8s-service/go.mod
@@ -1,4 +1,4 @@
-module github.com/42core-team/website_relaunch/k8s-service-gen
+module github.com/42core-team/website_relaunch/k8s-service
 
 go 1.24.2
