+++ k8s-service/internal/config/config.go
@@ -15,6 +15,8 @@ type Config struct {
 	RabbitMQ      string  `env:"RABBITMQ, default=amqp://guest:guest@localhost:5672/"`
 	RabbitMQHTTP  string  `env:"RABBITMQ_HTTP, default=http://localhost:15672"`
 	S3Endpoint    string  `env:"S3_ENDPOINT, required"`
+	S3Region      string  `env:"S3_REGION, default=eu"`
+	S3Bucket      string  `env:"S3_BUCKET, required"`
 	S3AccessKeyID string  `env:"S3_ACCESS_KEY_ID, required"`
 	S3SecretKey   string  `env:"S3_SECRET_ACCESS_KEY, required"`
 }
