+++ k8s-service/internal/queue/parse.go
@@ -3,7 +3,7 @@ package queue
 import (
 	"encoding/json"
 
-	"github.com/42core-team/website_relaunch/k8s-service-gen/internal/kube"
+	"github.com/42core-team/website_relaunch/k8s-service/internal/kube"
 )
 
 func parseGameMessage(msg []byte) (kube.Game, error) {
